Crate browser_window_c

Structs

c_G_fpos64_t

c_G_fpos_t

c_IO_FILE

c_IO_codecvt

c_IO_marker

c_IO_wide_data

c__fsid_t

c__locale_data

c__locale_struct

c__mbstate_t

c__pthread_cond_s

c__pthread_cond_s__bindgen_ty_1__bindgen_ty_1

c__pthread_cond_s__bindgen_ty_2__bindgen_ty_1

c__pthread_internal_list

c__pthread_internal_slist

c__pthread_mutex_s

c__pthread_rwlock_arch_t

c__sigset_t

c__va_list_tag

cbw_Application

cbw_ApplicationDispatchData

cbw_ApplicationEngineData

cbw_ApplicationEngineImpl

cbw_ApplicationImpl

cbw_ApplicationSettings

cbw_BrowserWindow

cbw_BrowserWindowImpl

cbw_BrowserWindowOptions

cbw_BrowserWindowSource

cbw_CStrSlice
A 'const string slice' Points to a immutable, non-zero-terminated, UTF-8 encoded string. Using rust's string's memory layout.

cbw_Dims2D

cbw_Err

cbw_Pos2D

cbw_StrSlice
A 'string slice' Points to a mutable, non-zero-terminated, UTF-8 encoded string. Using rust's string's memory layout.

cbw_Window

cbw_WindowCallbacks

cbw_WindowDispatchData

cbw_WindowImpl

cbw_WindowOptions

cdiv_t

cdrand48_data

cfd_set

cldiv_t

clldiv_t

cmax_align_t

crandom_data

ctimespec

ctimeval

Constants

cBIG_ENDIAN

cBUFSIZ

cBW_ERR_CODE_SUCCESS

cBYTE_ORDER

cE2BIG

cEACCES

cEADDRINUSE

cEADDRNOTAVAIL

cEADV

cEAFNOSUPPORT

cEAGAIN

cEALREADY

cEBADE

cEBADF

cEBADFD

cEBADMSG

cEBADR

cEBADRQC

cEBADSLT

cEBFONT

cEBUSY

cECANCELED

cECHILD

cECHRNG

cECOMM

cECONNABORTED

cECONNREFUSED

cECONNRESET

cEDEADLK

cEDEADLOCK

cEDESTADDRREQ

cEDOM

cEDOTDOT

cEDQUOT

cEEXIST

cEFAULT

cEFBIG

cEHOSTDOWN

cEHOSTUNREACH

cEHWPOISON

cEIDRM

cEILSEQ

cEINPROGRESS

cEINTR

cEINVAL

cEIO

cEISCONN

cEISDIR

cEISNAM

cEKEYEXPIRED

cEKEYREJECTED

cEKEYREVOKED

cEL2HLT

cEL2NSYNC

cEL3HLT

cEL3RST

cELIBACC

cELIBBAD

cELIBEXEC

cELIBMAX

cELIBSCN

cELNRNG

cELOOP

cEMEDIUMTYPE

cEMFILE

cEMLINK

cEMSGSIZE

cEMULTIHOP

cENAMETOOLONG

cENAVAIL

cENETDOWN

cENETRESET

cENETUNREACH

cENFILE

cENOANO

cENOBUFS

cENOCSI

cENODATA

cENODEV

cENOENT

cENOEXEC

cENOKEY

cENOLCK

cENOLINK

cENOMEDIUM

cENOMEM

cENOMSG

cENONET

cENOPKG

cENOPROTOOPT

cENOSPC

cENOSR

cENOSTR

cENOSYS

cENOTBLK

cENOTCONN

cENOTDIR

cENOTEMPTY

cENOTNAM

cENOTRECOVERABLE

cENOTSOCK

cENOTSUP

cENOTTY

cENOTUNIQ

cENXIO

cEOF

cEOPNOTSUPP

cEOVERFLOW

cEOWNERDEAD

cEPERM

cEPFNOSUPPORT

cEPIPE

cEPROTO

cEPROTONOSUPPORT

cEPROTOTYPE

cERANGE

cEREMCHG

cEREMOTE

cEREMOTEIO

cERESTART

cERFKILL

cEROFS

cESHUTDOWN

cESOCKTNOSUPPORT

cESPIPE

cESRCH

cESRMNT

cESTALE

cESTRPIPE

cETIME

cETIMEDOUT

cETOOMANYREFS

cETXTBSY

cEUCLEAN

cEUNATCH

cEUSERS

cEWOULDBLOCK

cEXDEV

cEXFULL

cEXIT_FAILURE

cEXIT_SUCCESS

cFALSE

cFD_SETSIZE

cFILENAME_MAX

cFOPEN_MAX

cINT8_MAX

cINT8_MIN

cINT16_MAX

cINT16_MIN

cINT32_MAX

cINT32_MIN

cINTPTR_MAX

cINTPTR_MIN

cINT_FAST8_MAX

cINT_FAST8_MIN

cINT_FAST16_MAX

cINT_FAST16_MIN

cINT_FAST32_MAX

cINT_FAST32_MIN

cINT_LEAST8_MAX

cINT_LEAST8_MIN

cINT_LEAST16_MAX

cINT_LEAST16_MIN

cINT_LEAST32_MAX

cINT_LEAST32_MIN

cLITTLE_ENDIAN

cL_ctermid

cL_tmpnam

cPDP_ENDIAN

cPTRDIFF_MAX

cPTRDIFF_MIN

cP_tmpdir

cRAND_MAX

cSEEK_CUR

cSEEK_END

cSEEK_SET

cSIG_ATOMIC_MAX

cSIG_ATOMIC_MIN

cSIZE_MAX

cTMP_MAX

cTRUE

cUINT8_MAX

cUINT16_MAX

cUINT32_MAX

cUINTPTR_MAX

cUINT_FAST8_MAX

cUINT_FAST16_MAX

cUINT_FAST32_MAX

cUINT_LEAST8_MAX

cUINT_LEAST16_MAX

cUINT_LEAST32_MAX

cWCONTINUED

cWEXITED

cWINT_MAX

cWINT_MIN

cWNOHANG

cWNOWAIT

cWSTOPPED

cWUNTRACED

c_ALLOCA_H

c_ASSERT_H

c_ATFILE_SOURCE

c_BITS_BYTESWAP_H

c_BITS_ENDIANNESS_H

c_BITS_ENDIAN_H

c_BITS_ERRNO_H

c_BITS_PTHREADTYPES_ARCH_H

c_BITS_PTHREADTYPES_COMMON_H

c_BITS_STDINT_INTN_H

c_BITS_STDINT_UINTN_H

c_BITS_STDIO_LIM_H

c_BITS_TIME64_H

c_BITS_TYPESIZES_H

c_BITS_TYPES_H

c_BITS_TYPES_LOCALE_T_H

c_BITS_TYPES___LOCALE_T_H

c_BITS_UINTN_IDENTITY_H

c_BITS_WCHAR_H

c_DEFAULT_SOURCE

c_ENDIAN_H

c_ERRNO_H

c_FEATURES_H

c_IOFBF

c_IOLBF

c_IONBF

c_IO_EOF_SEEN

c_IO_ERR_SEEN

c_IO_USER_LOCK

c_POSIX_C_SOURCE

c_POSIX_SOURCE

c_STDC_PREDEF_H

c_STDINT_H

c_STDIO_H

c_STDLIB_H

c_STRINGS_H

c_STRING_H

c_STRUCT_TIMESPEC

c_SYS_CDEFS_H

c_SYS_SELECT_H

c_SYS_TYPES_H

c_THREAD_MUTEX_INTERNAL_H

c_THREAD_SHARED_TYPES_H

c__BIG_ENDIAN

c__BIT_TYPES_DEFINED__

c__BYTE_ORDER

c__ENUM_IDTYPE_T

c__FD_SETSIZE

c__FD_ZERO_STOS

c__FILE_defined

c__FLOAT_WORD_ORDER

c__GLIBC_MINOR__

c__GLIBC_USE_DEPRECATED_GETS

c__GLIBC_USE_DEPRECATED_SCANF

c__GLIBC_USE_IEC_60559_BFP_EXT

c__GLIBC_USE_IEC_60559_BFP_EXT_C2X

c__GLIBC_USE_IEC_60559_FUNCS_EXT

c__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

c__GLIBC_USE_IEC_60559_TYPES_EXT

c__GLIBC_USE_ISOC2X

c__GLIBC_USE_LIB_EXT2

c__GLIBC__

c__GNUC_VA_LIST

c__GNU_LIBRARY__

c__HAVE_DISTINCT_FLOAT16

c__HAVE_DISTINCT_FLOAT32

c__HAVE_DISTINCT_FLOAT32X

c__HAVE_DISTINCT_FLOAT64

c__HAVE_DISTINCT_FLOAT64X

c__HAVE_DISTINCT_FLOAT128

c__HAVE_DISTINCT_FLOAT128X

c__HAVE_FLOAT16

c__HAVE_FLOAT32

c__HAVE_FLOAT32X

c__HAVE_FLOAT64

c__HAVE_FLOAT64X

c__HAVE_FLOAT64X_LONG_DOUBLE

c__HAVE_FLOAT128

c__HAVE_FLOAT128X

c__HAVE_FLOATN_NOT_TYPEDEF

c__HAVE_GENERIC_SELECTION

c__INO_T_MATCHES_INO64_T

c__LITTLE_ENDIAN

c__LONG_DOUBLE_USES_FLOAT128

c__OFF_T_MATCHES_OFF64_T

c__PDP_ENDIAN

c__PTHREAD_MUTEX_HAVE_PREV

c__RLIM_T_MATCHES_RLIM64_T

c__SIZEOF_PTHREAD_ATTR_T

c__SIZEOF_PTHREAD_BARRIERATTR_T

c__SIZEOF_PTHREAD_BARRIER_T

c__SIZEOF_PTHREAD_CONDATTR_T

c__SIZEOF_PTHREAD_COND_T

c__SIZEOF_PTHREAD_MUTEXATTR_T

c__SIZEOF_PTHREAD_MUTEX_T

c__SIZEOF_PTHREAD_RWLOCKATTR_T

c__SIZEOF_PTHREAD_RWLOCK_T

c__STATFS_MATCHES_STATFS64

c__STDC_IEC_559_COMPLEX__

c__STDC_IEC_559__

c__STDC_ISO_10646__

c__SYSCALL_WORDSIZE

c__TIMESIZE

c__USE_ATFILE

c__USE_FORTIFY_LEVEL

c__USE_ISOC11

c__USE_ISOC95

c__USE_ISOC99

c__USE_MISC

c__USE_POSIX

c__USE_POSIX2

c__USE_POSIX199309

c__USE_POSIX199506

c__USE_POSIX_IMPLICITLY

c__USE_XOPEN2K

c__USE_XOPEN2K8

c__WALL

c__WCLONE

c__WCOREFLAG

c__WNOTHREAD

c__WORDSIZE

c__WORDSIZE_TIME64_COMPAT32

c__W_CONTINUED

c____FILE_defined

c_____fpos64_t_defined

c_____fpos_t_defined

c____mbstate_t_defined

c__bool_true_false_are_defined

c__clock_t_defined

c__clockid_t_defined

c__glibc_c99_flexarr_available

c__have_pthread_attr_t

c__ldiv_t_defined

c__lldiv_t_defined

c__sigset_t_defined

c__struct_FILE_defined

c__time_t_defined

c__timer_t_defined

c__timeval_defined

cfalse

cidtype_t_P_ALL

cidtype_t_P_PGID

cidtype_t_P_PID

ctrue

Statics

cstderr

cstdin

cstdout

csys_errlist

csys_nerr

Functions

c_Exit⚠

c__assert⚠

c__assert_fail⚠

c__assert_perror_fail⚠

c__ctype_get_mb_cur_max⚠

c__errno_location⚠

c__getdelim⚠

c__overflow⚠

c__stpcpy⚠

c__stpncpy⚠

c__strtok_r⚠

c__uflow⚠

c_bw_Window_onResize⚠

ca64l⚠

cabort⚠

cabs⚠

caligned_alloc⚠

calloca⚠

cat_quick_exit⚠

catexit⚠

catof⚠

catoi⚠

catol⚠

catoll⚠

cbcmp⚠

cbcopy⚠

cbsearch⚠

cbw_Application_assertCorrectThread⚠
Safety check that makes sure the given application handle is used on the correct thread. Does nothing in release mode.

cbw_Application_dispatch⚠
Dispatches the given function to be executed on the thread this application instance has been created on, and passes the given data to it. This function is thread safe.

cbw_Application_exit⚠
Exits the main loop, returning execution to the function that invoked the run call. The exit_code will be returned by bw_Application_run.

cbw_Application_exitAsync⚠
Same as bw_Application_exit, but guaranteed to be thread-safe The exit_code will be returned by bw_Application_run.

cbw_Application_finish⚠
Shuts down all application processes and performs necessary clean-up code.

cbw_Application_free⚠
Frees memory for the application handle. Call bw_Application_finish before you call this function. Freeing the application handle is generally not necessary, as all memory in use by the process gets released anyway after shutdown.

cbw_Application_initialize⚠
Initializes browser window. Starts up browser engine process(es). Returns an application handle.

cbw_Application_isRunning⚠

cbw_Application_run⚠
Runs the event loop. Calls the on_ready callback when app can be used.

cbw_BrowserWindow_destroy⚠

cbw_BrowserWindow_drop⚠
Marks the browser window handle as not being used anymore. This makes it so that if and when the window gets closed, everything is freed and cleaned from memory. This function can be called from any thread so it needs to be thread safe.

cbw_BrowserWindow_evalJs⚠
Executes the given JavaScript and calls the given callback (on the GUI thread) to provide the result.

cbw_BrowserWindow_evalJsThreaded⚠

cbw_BrowserWindow_getApp⚠

cbw_BrowserWindow_getUserData⚠

cbw_BrowserWindow_getWindow⚠

cbw_BrowserWindow_navigate⚠

cbw_BrowserWindow_new⚠
Creates a new browser window

cbw_Err_free⚠

cbw_Err_msg_string⚠

cbw_Err_msg_success⚠

cbw_Err_new_with_msg⚠

cbw_Window_destroy⚠
Destroy the window, releasing all resources it holds. After this call, the handle is considered invalid.

cbw_Window_drop⚠
Invalidates the window handle. The window will get destroyed when it is deemed possible.

cbw_Window_getContentDimensions⚠
Gets the width and height of the usable area inside the window.

cbw_Window_getOpacity⚠
Gets the opacity of the window as a value from 0 to 255.

cbw_Window_getPosition⚠
Gets the X and Y coordinates of the window position relative to the desktop screen.

cbw_Window_getTitle⚠
Copies as many bytes into title that fit in there. Returns the number of characters the title actually has.

cbw_Window_getWindowDimensions⚠
Gets the width and height of the window including the title bar and borders.

cbw_Window_hide⚠

cbw_Window_isVisible⚠
Returns whether or not the window is not hidden. bw_Window_show and bw_Window_hide change the visibility.

cbw_Window_new⚠
Creates a new (empty) window The returned pointer is a handler for the window. bw_Window_drop needs to be called on it after it is done being used, otherwise the window is never actually destroyed and memory leakes happen.

cbw_Window_setContentDimensions⚠

cbw_Window_setOpacity⚠

cbw_Window_setPosition⚠

cbw_Window_setTitle⚠
Applies the given title;

cbw_Window_setWindowDimensions⚠

cbw_Window_show⚠
Shows the window if it was previously hidden Is generally called after window creation.

cbw_Window_triggerClose⚠

cbw_string_copyAsNewCstr⚠
Copies the string from the given bw_CStrSlice to a C compatible, nul-terminated string.

cbw_string_freeCstr⚠
Frees the string allocated with any of the functions of this module.

cbzero⚠

ccalloc⚠

cclearenv⚠

cclearerr⚠

cclearerr_unlocked⚠

cctermid⚠

cdiv⚠

cdprintf⚠

cdrand48⚠

cdrand48_r⚠

cecvt⚠

cecvt_r⚠

cerand48⚠

cerand48_r⚠

cexit⚠

cexplicit_bzero⚠

cfclose⚠

cfcvt⚠

cfcvt_r⚠

cfdopen⚠

cfeof⚠

cfeof_unlocked⚠

cferror⚠

cferror_unlocked⚠

cfflush⚠

cfflush_unlocked⚠

cffs⚠

cffsl⚠

cffsll⚠

cfgetc⚠

cfgetc_unlocked⚠

cfgetpos⚠

cfgets⚠

cfileno⚠

cfileno_unlocked⚠

cflockfile⚠

cfmemopen⚠

cfopen⚠

cfprintf⚠

cfputc⚠

cfputc_unlocked⚠

cfputs⚠

cfread⚠

cfread_unlocked⚠

cfree⚠

cfreopen⚠

cfscanf⚠

cfscanf1⚠

cfseek⚠

cfseeko⚠

cfsetpos⚠

cftell⚠

cftello⚠

cftrylockfile⚠

cfunlockfile⚠

cfwrite⚠

cfwrite_unlocked⚠

cgcvt⚠

cgetc⚠

cgetc_unlocked⚠

cgetchar⚠

cgetchar_unlocked⚠

cgetdelim⚠

cgetenv⚠

cgetline⚠

cgetloadavg⚠

cgetsubopt⚠

cgetw⚠

cindex⚠

cinitstate⚠

cinitstate_r⚠

cjrand48⚠

cjrand48_r⚠

cl64a⚠

clabs⚠

clcong48⚠

clcong48_r⚠

cldiv⚠

cllabs⚠

clldiv⚠

clrand48⚠

clrand48_r⚠

cmalloc⚠

cmblen⚠

cmbstowcs⚠

cmbtowc⚠

cmemccpy⚠

cmemchr⚠

cmemcmp⚠

cmemcpy⚠

cmemmove⚠

cmemset⚠

cmkdtemp⚠

cmkstemp⚠

cmkstemps⚠

cmktemp⚠

cmrand48⚠

cmrand48_r⚠

cnrand48⚠

cnrand48_r⚠

con_exit⚠

copen_memstream⚠

cpclose⚠

cperror⚠

cpopen⚠

cposix_memalign⚠

cprintf⚠

cpselect⚠

cputc⚠

cputc_unlocked⚠

cputchar⚠

cputchar_unlocked⚠

cputenv⚠

cputs⚠

cputw⚠

cqecvt⚠

cqecvt_r⚠

cqfcvt⚠

cqfcvt_r⚠

cqgcvt⚠

cqsort⚠

cquick_exit⚠

crand⚠

crand_r⚠

crandom⚠

crandom_r⚠

crealloc⚠

creallocarray⚠

crealpath⚠

cremove⚠

crename⚠

crenameat⚠

crewind⚠

crindex⚠

crpmatch⚠

cscanf⚠

cscanf1⚠

cseed48⚠

cseed48_r⚠

cselect⚠

csetbuf⚠

csetbuffer⚠

csetenv⚠

csetlinebuf⚠

csetstate⚠

csetstate_r⚠

csetvbuf⚠

csnprintf⚠

csprintf⚠

csrand⚠

csrand48⚠

csrand48_r⚠

csrandom⚠

csrandom_r⚠

csscanf⚠

csscanf1⚠

cstpcpy⚠

cstpncpy⚠

cstrcasecmp⚠

cstrcasecmp_l⚠

cstrcat⚠

cstrchr⚠

cstrcmp⚠

cstrcoll⚠

cstrcoll_l⚠

cstrcpy⚠

cstrcspn⚠

cstrdup⚠

cstrerror⚠

cstrerror_l⚠

cstrerror_r⚠

cstrlen⚠

cstrncasecmp⚠

cstrncasecmp_l⚠

cstrncat⚠

cstrncmp⚠

cstrncpy⚠

cstrndup⚠

cstrnlen⚠

cstrpbrk⚠

cstrrchr⚠

cstrsep⚠

cstrsignal⚠

cstrspn⚠

cstrstr⚠

cstrtod⚠

cstrtof⚠

cstrtok⚠

cstrtok_r⚠

cstrtol⚠

cstrtold⚠

cstrtoll⚠

cstrtoq⚠

cstrtoul⚠

cstrtoull⚠

cstrtouq⚠

cstrxfrm⚠

cstrxfrm_l⚠

csystem⚠

ctempnam⚠

ctmpfile⚠

ctmpnam⚠

ctmpnam_r⚠

cungetc⚠

cunsetenv⚠

cvalloc⚠

cvdprintf⚠

cvfprintf⚠

cvfscanf⚠

cvfscanf1⚠

cvprintf⚠

cvscanf⚠

cvscanf1⚠

cvsnprintf⚠

cvsprintf⚠

cvsscanf⚠

cvsscanf1⚠

cwcstombs⚠

cwctomb⚠

Type Definitions

cBOOL

cFILE

c_Float32

c_Float32x

c_Float64

c_Float64x

c_IO_lock_t

c__FILE

c__blkcnt64_t

c__blkcnt_t

c__blksize_t

c__builtin_va_list

c__caddr_t

c__clock_t

c__clockid_t

c__compar_fn_t

c__daddr_t

c__dev_t

c__fd_mask

c__fpos64_t

c__fpos_t

c__fsblkcnt64_t

c__fsblkcnt_t

c__fsfilcnt64_t

c__fsfilcnt_t

c__fsword_t

c__gid_t

c__gnuc_va_list

c__id_t

c__ino64_t

c__ino_t

c__int8_t

c__int16_t

c__int32_t

c__int64_t

c__int_least8_t

c__int_least16_t

c__int_least32_t

c__int_least64_t

c__intmax_t

c__intptr_t

c__key_t

c__locale_t

c__loff_t

c__mode_t

c__nlink_t

c__off64_t

c__off_t

c__pid_t

c__pthread_list_t

c__pthread_slist_t

c__quad_t

c__rlim64_t

c__rlim_t

c__sig_atomic_t

c__socklen_t

c__ssize_t

c__suseconds_t

c__syscall_slong_t

c__syscall_ulong_t

c__time_t

c__timer_t

c__u_char

c__u_int

c__u_long

c__u_quad_t

c__u_short

c__uid_t

c__uint8_t

c__uint16_t

c__uint32_t

c__uint64_t

c__uint_least8_t

c__uint_least16_t

c__uint_least32_t

c__uint_least64_t

c__uintmax_t

c__useconds_t

cblkcnt_t

cblksize_t

cbw_ApplicationDispatchFn

cbw_ApplicationReadyFn

cbw_BrowserWindowCreationCallbackFn

cbw_BrowserWindowHandlerFn

cbw_BrowserWindowJsCallbackFn

cbw_ErrCode

cbw_WindowDispatchFn

ccaddr_t

cclock_t

cclockid_t

cdaddr_t

cdev_t

cfd_mask

cfpos_t

cfsblkcnt_t

cfsfilcnt_t

cfsid_t

cgid_t

cid_t

cidtype_t

cino_t

cint_fast8_t

cint_fast16_t

cint_fast32_t

cint_fast64_t

cint_least8_t

cint_least16_t

cint_least32_t

cint_least64_t

cintmax_t

ckey_t

clocale_t

cloff_t

cmode_t

cnlink_t

coff_t

cpid_t

cpthread_key_t

cpthread_once_t

cpthread_spinlock_t

cpthread_t

cquad_t

cregister_t

csigset_t

csize_t

cssize_t

csuseconds_t

ctime_t

ctimer_t

cu_char

cu_int

cu_int8_t

cu_int16_t

cu_int32_t

cu_int64_t

cu_long

cu_quad_t

cu_short

cuid_t

cuint

cuint_fast8_t

cuint_fast16_t

cuint_fast32_t

cuint_fast64_t

cuint_least8_t

cuint_least16_t

cuint_least32_t

cuint_least64_t

cuintmax_t

culong

cushort

cva_list

cwchar_t

Unions

c__mbstate_t__bindgen_ty_1

c__pthread_cond_s__bindgen_ty_1

c__pthread_cond_s__bindgen_ty_2

cpthread_attr_t

cpthread_barrier_t

cpthread_barrierattr_t

cpthread_cond_t

cpthread_condattr_t

cpthread_mutex_t

cpthread_mutexattr_t

cpthread_rwlock_t

cpthread_rwlockattr_t